auth-service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116",
    "uvicorn[standard]>=0.35",
    "sqlalchemy>=2.0",
    "alembic>=1.16",
    "pydantic>=2.11",
    "pydantic-settings>=2.10",
    "python-dotenv>=1.1",
    "python-jose[cryptography]>=3.5",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9",
    "supabase>=2.17",
    "sendgrid>=6.12",
    "slowapi>=0.1.9",
    "email-validator>=2.2",
    "prometheus-fastapi-instrumentator>=6.1",
    "google-auth>=2.40",
    "cachecontrol>=0.14",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=70"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# E/W = pycodestyle, F = pyflakes, I = import order, UP = pyupgrade (modern Python syntax)
select = ["E", "F", "I", "W", "UP"]
ignore = [
    "E501",  # line too long — formatter handles this
]
